Use base64 hash instead of sha256 hash in lambda_function
When using sha256 hash, terraform/AWS converts it to base64, and then
stores it. This poses a problem next time terraform apply is ran, as the
lambda function's source_code_hash is set to the base64 version of the
hash, while our hash is the sha256 hash. This forced an update of the
source code every time terraform apply ran.

This change fixes this problem.