Move path handling logic out of lambda handler
Lambda cannot reuse the database connection if it is in the handler, it
has to be globally scoped. If it is globally scoped the tests will fail
as they mock the database connection.

This moves the 'business' logic outside of the handler so that it can be
tested separately.

Part of issue #1.