Expose a way to specify a role terraform can assume
By adding this, the credentials that terraform is called with no longer
need to have access to all resources. Instead they are only required to
have access to the role that provides that access. -> Better security

To ahve backward compatibility with current use (where the credentials
that call terraform have access to all resources), no role is assumed
when the default terraform workspace is used.